correções no inimigo.cpp e inimigo.hpp, adição do combate.cpp e combate.hpp, substituição do gameloop.cpp por main.cpp

+++ inimigo.cpp
@@ -1,7 +1,7 @@
 #include <string>
 #include "inimigo.hpp"
 
-Inimigo::Inimigo(std::string nome, int vida, int dano){
+Inimigo::Inimigo(std::string nome, double vida, double dano){
     this -> nome = nome;
     this -> vida = vida;
     this -> dano = dano;
